mayankPbl/ocr/requirements.txt
pdfplumber>=0.10.0

# Revenue Agent dependencies
numpy>=1.26
pandas>=2.2,<3.0
langchain>=0.2
langchain-openai>=0.1

# Sentiment Agent — NewsAPI HTTP client
requests>=2.28

# yfinance ingestion — fetch listed company financials by ticker
yfinance>=0.2

# Private company ingestion — CSV / Excel uploads
openpyxl>=3.1

# Dashboard
streamlit>=1.34
streamlit-agraph>=0.0.45
graphviz>=0.20

# ── Auth & Database ───────────────────────────────────────────────────────────
pymongo>=4.6            # MongoDB driver
bcrypt>=4.1             # Password hashing
PyJWT>=2.8              # JWT token creation / validation
python-dotenv>=1.0      # Load .env into os.environ
dnspython>=2.4          # Required for MongoDB Atlas SRV connection strings
